move request validation and status page to http config
